+++ mitra_cli/src/commands/activitypub.rs
@@ -28,6 +28,8 @@ use mitra_activitypub::{
         like::build_like,
     },
     deliverer::{Recipient, Sender},
+    forwarder::get_activity_recipients,
+    handlers::activity::get_activity_audience,
     identifiers::canonicalize_id,
     importers::{
         get_user_by_actor_id,
@@ -419,7 +421,8 @@ pub struct SendActivity {
     /// JSON value
     activity: String,
     /// Actor ID
-    recipient: String,
+    #[arg(long)]
+    recipient: Option<String>,
     /// Create RFC-9421 signature?
     #[arg(long)]
     rfc9421: bool,
@@ -442,10 +445,23 @@ impl SendActivity {
             &authority,
             &canonical_actor_id,
         ).await?;
-        let recipient = get_remote_profile_by_actor_id(
-            db_client_await!(db_pool),
-            &self.recipient,
-        ).await?;
+
+        let recipient = if let Some(recipient) = self.recipient {
+            get_remote_profile_by_actor_id(
+                db_client_await!(db_pool),
+                &recipient,
+            ).await?
+        } else {
+            let audience = get_activity_audience(&activity, None)?;
+            let recipients = get_activity_recipients(
+                db_client_await!(db_pool),
+                &audience,
+            ).await?;
+            recipients
+                .into_iter()
+                .next()
+                .ok_or(Error::msg("recipient can not be determined"))?
+        };
         let recipient_inbox = Recipient
             ::for_inbox(recipient.expect_actor_data())
             .first()